requirements.txt
Flask==3.0.3
requests==2.32.3
pillow==10.3.0
numpy==2.0.0
joblib==1.4.2
scikit-learn==1.5.0